% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babelnet_to_wordnet.R
\name{babelnet_to_wordnet}
\alias{babelnet_to_wordnet}
\title{BabelNet to WordNet}
\usage{
babelnet_to_wordnet(ids, outFile = NULL, cache = paste0(dirname(tempdir()),
  "/babelnet_sparql/"), wordnet_dir = "~/WordNet-3.0",
  download_wordnet = FALSE, wordnet_version = 30)
}
\arguments{
\item{ids}{A vector of BabelNet IDs (e.g., \code{"bn:00103811a"}).}

\item{outFile}{Path of a file to save results to.}

\item{cache}{Path to a directory in which to save original BabelNet and WordNet files.}

\item{wordnet_dir}{Path to a WordNet directory (the unpacked tars), ultimately
containing a \code{/dict/index.sense} file.}

\item{download_wordnet}{Logical; if \code{TRUE}, will download the 3.0 version of WordNet,
if \code{wordnet_dir} does not exist.}

\item{wordnet_version}{Version of WordNet pointed to by \code{wordnet_dir},
which determines which \code{old_keys} entry is looked for (e.g., \code{30} for version 3.0).}
}
\value{
A \code{data.frame} with a map between provided \code{ids} and WordNet information.
}
\description{
Retrieve \href{https://wordnet.princeton.edu}{WordNet} counterparts to
\href{https://babelnet.org/}{BabelNet} IDs where available.
}
\examples{
\dontrun{
babelnet_to_wordnet("bn:00103811a")
}
}
