% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babelnet_synset.R
\name{bablenet_synset}
\alias{bablenet_synset}
\title{BabelNet Synset}
\usage{
bablenet_synset(ids, outDir = NULL, overwrite = FALSE,
  cache = paste0(dirname(tempdir()), "/babelnet_pages/"),
  cores = parallel::detectCores() - 2)
}
\arguments{
\item{ids}{A vector of BabelNet IDs (e.g., \code{"bn:00103811a"}).}

\item{outDir}{Path of a directory in which to save parsed page JSON files.}

\item{overwrite}{Logical; if \code{TRUE}, will overwrite existing parsed pages when retrieved.}

\item{cache}{Path to a directory in which to save original page HTML files.}

\item{cores}{Number of CPU cores to use during processing}
}
\value{
A list with entries for each id in \code{ids}.
}
\description{
Retrieve \href{https://babelnet.org/}{BabelNet} synset information.
}
\examples{
bablenet_synset("bn:00103811a")
}
