% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspmed.R
\name{bspmed}
\alias{bspmed}
\title{Bootstrapped / Permuted Mediation}
\usage{
bspmed(x, y, m, data = NULL, cov = NULL, random = NULL, su = NA,
  i = 1000, method = "permutation", alpha = 0.05, figure = TRUE, ...)
}
\arguments{
\item{x}{X variable: can be the name of a variable in the environment or data.}

\item{y}{Y variable: can be the name of a variable in the environment or data.}

\item{m}{Mediating variable: can be the name of a variable in the environment or data.}

\item{data}{\code{data.frame} from which to pull variable, with the global environment as a fallback.}

\item{cov}{A character vector of covariates to include in each path.
Only accepted if \code{data} is also specified.}

\item{random}{Variable to be included as a random intercept in an \code{\link[nlme]{lme}} model:
\code{lme(y ~ x + m, random = ~ 1 | random)}}

\item{su}{Subset applied across variables, e.g., \code{su = x < 2 & m > 0}.}

\item{i}{The number of iterations to be processed for bootstrapping. Default is \code{1000}.}

\item{method}{Character specifying the test type. Anything starting with \code{b} will perform
bootstrapping, otherwise will performs the non-iterative, permutation-of-residuals approach as
described in
\href{https://link.springer.com/article/10.3758/s13428-011-0181-x}{Taylor and MacKinnon (2012; Eqs. 9 and 10)}.}

\item{alpha}{The alpha level for the bootstrapping confidence interval.
Default is \code{.05} (i.e., 2.5\% and 97.5\%).}

\item{figure}{Logical; if \code{FALSE}, will not display a figure.}

\item{...}{Passes additional arguments to \code{\link{medfig}} if \code{figure} is \code{TRUE}.}
}
\description{
Displays the results of Sobel and permutation/bootstrapping tests for simple mediation.
}
\details{
Bootstrap confidence intervals indicate significance when they do not include zero (as they are
confidence intervals of the effect). Permutation confidence intervals indicate significance when
they do not include the indirect effect (as they are confidence intervals of the null effect).
}
\examples{
# This displays the mediation of weight on the relationship
# between displacement and miles per gallon.
bspmed(disp, mpg, wt, scale(mtcars))
}
\references{
Taylor, A. B., & MacKinnon, D. P. (2012). Four applications of permutation methods to testing a
single-mediator model. \emph{Behavior research methods, 44}, 806-844. doi:
\href{https://doi.org/10.3758/s13428-011-0181-x}{10.3758/s13428-011-0181-x}
}
\seealso{
\code{\link{medfig}} to make a figure from results.
}
