% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_baby_names.R
\name{get_baby_names}
\alias{get_baby_names}
\title{Get Baby Names Data}
\usage{
get_baby_names(dir = tempdir(), source = "national")
}
\arguments{
\item{dir}{Directory in which to save data.}

\item{source}{Which dataset to download; between \code{national} (default), \code{state}, and \code{territory}.}
}
\value{
A list with entries for \code{data} (a \code{data.frame} with the original data), and
\code{summary} (a \code{matrix} with a count of each name and probabilities per sex).
}
\description{
Downloads and processes historical baby names data from the U.S.
\href{https://www.ssa.gov/OACT/babynames/limits.html}{Social Security Administration}.
}
\examples{
\dontrun{
# saves to a temporary directory
names <- get_baby_names()
}
}
