% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_sex.R
\name{guess_sex}
\alias{guess_sex}
\title{Guess sex based on a web search}
\usage{
guess_sex(names, retry = 50,
  search_source = "https://search.yahoo.com/search?q=",
  cores = detectCores() - 2)
}
\arguments{
\item{names}{A character vector of search terms (such as a name).}

\item{retry}{Logical; if \code{TRUE}, will retry after a failed attempt.}

\item{search_source}{URL of the search engine; each entry in \code{names}
is simply appended.}

\item{cores}{Number of CPU cores to split \code{names} across.}
}
\value{
a \code{data.frame} with entries for \code{name} (the original search term),
\code{guess}, \code{confidence}, and \code{female} and \code{male} with count of matches.
}
\description{
Searches for a name (or whatever), and looks for gendered pronouns or other basic cues to sex.
}
\examples{
\dontrun{
guess_sex(c("Jane", "John"))
}
}
