% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lusilab-package.R
\docType{package}
\name{lusilab-package}
\alias{lusilab}
\alias{lusilab-package}
\title{lusilab: Language Use and Social Interaction Lab}
\description{
Includes functions for power analysis, permuting or bootstrapping mediations, automated factor analysis, and pulling text from reddit, imsdb, and rottentomatoes, plus some other things to play with.
}
\keyword{internal}
