% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personality_database.R
\name{personality_database}
\alias{personality_database}
\title{Personality-Database API}
\usage{
personality_database(profiles = "famous", category = "all",
  outFile = NULL, sort = "alphabet", start = 0, limit = Inf,
  retry = 10, sleep_time = 60, verbose = TRUE,
  cache = paste0(tempdir(), "/personality_database/"), list_cache = NULL,
  cores = detectCores() - 1)
}
\arguments{
\item{profiles}{Name of the profile list, between \code{famous}, \code{fictional}, and \code{discovery}.}

\item{category}{Name of the profile category, as listed under each list, or \code{all} (default).}

\item{outFile}{Path to a .csv file to save results as.}

\item{sort}{How the profile list should be sorted, between \code{alphabet} (default), \code{hot}, \code{new}, or \code{top}.}

\item{start}{Profile from the list to start on.}

\item{limit}{Maximum number of profiles to include.}

\item{retry}{Logical; if \code{TRUE}, will retry after a failed attempt.}

\item{sleep_time}{Number of seconds to wait between retries.}

\item{verbose}{Logical; if \code{FALSE}, will not display status messages.}

\item{cache}{Path to a directory in which to save raw responses.}

\item{list_cache}{A saved profile list, or path to such a file, to collect scores for. Providing this
avoids collecting a list initially.}

\item{cores}{Number of CPU cores to split \code{names} across.}
}
\value{
a \code{data.frame} with entries for \code{name} (the original search term),
\code{guess}, \code{confidence}, and \code{female} and \code{male} with count of matches.
}
\description{
Collect data from the \href{https://www.personality-database.com}{personality-database} API.
}
\examples{
\dontrun{
# get scores from the 5 "HOT"est famous people
hot5 <- personality_database(limit = 5, sort = "hot")
}
}
