% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_demographics.R
\name{predict_demographics}
\alias{predict_demographics}
\title{Predict Demographics From Name}
\usage{
predict_demographics(given, family = NULL, country = NULL,
  source = "all", dir = tempdir(), ssa_source = "national",
  full_country = FALSE, verbose = TRUE)
}
\arguments{
\item{given}{A vector of given (first) names. Case sensitive for all but \code{wgnd} and \code{fb_scraped}.}

\item{family}{A vector of family (last; sur) names. Only used in \code{fb} to adjust country predictions. Case sensitive.}

\item{country}{A vector of 2-letter country codes. Only used in \code{wgnd} to give \code{sex_in_country_wgnd}.}

\item{source}{A vector specify which source(s) to use:
\itemize{
\item \href{https://dataverse.harvard.edu/dataverse/WGND}{World Gender Name Dictionary} (\code{wgnd})
\item \href{https://github.com/philipperemy/name-dataset}{Facebook Hack} (\code{fb})
\item \href{https://sites.google.com/site/facebooknamelist/home}{Facebook Scraped} (\code{fb_scraped})
\item \href{https://archive.ics.uci.edu/ml/datasets/Gender+by+Name#}{Skydeck Censuses} (\code{skydeck})
\item \href{https://www.ssa.gov/OACT/babynames/limits.html}{U.S. Social Security Administration Baby Names} (\code{usssa})
}
Defaults to \code{all} to return all.}

\item{dir}{Directory in which to save original and prepared names data.}

\item{ssa_source}{Source of USSSA data; passed to \code{\link{get_baby_names}}.}

\item{full_country}{Logical; if \code{TRUE}, a column for each country will be returned for source
\code{fb}, each containing a probability for that country. Otherwise, only the most likely country
is returned in a \code{predicted_country} column.}

\item{verbose}{Logical; if \code{FALSE}, will not show status messages.}
}
\value{
A \code{data.frame} with columns for \code{given}, \code{family}, and \code{country} (as provided),
and columns for each \code{source}, including \code{count} and \code{prob_fem} prefixed by the source code,
and for \code{fb}, either \code{predicted_country_fb}, or a set of country codes prefixed by \code{_fb}.
}
\description{
Predict sex and maybe country from a name.
}
\examples{
\dontrun{

# saves to a temporary directory
predict_demographics(c("Jane", "John"), c("Doe", "Doe"))
}
}
