% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personality_database.R
\name{personality_database}
\alias{personality_database}
\title{Personality-Database API}
\usage{
personality_database(profiles = "famous", category = "all",
  outFile = NULL, sort = "hot", start = 0, limit = Inf, retry = 10,
  verbose = TRUE, cores = detectCores() - 1)
}
\arguments{
\item{profiles}{Name of the profile list, between \code{famous}, \code{fictional}, and \code{discovery}.}

\item{category}{Name of the profile category, as listed under each list, or \code{all} (default).}

\item{outFile}{Path to a .csv file to save results as.}

\item{sort}{How the profile list should be sorted, between \code{hot} (default), \code{new}, \code{top},
and \code{alphabet}.}

\item{start}{Profile from the list to start on.}

\item{limit}{Maximum number of profiles to include.}

\item{retry}{Logical; if \code{TRUE}, will retry after a failed attempt.}

\item{verbose}{Logical; if \code{FALSE}, will not display status messages.}

\item{cores}{Number of CPU cores to split \code{names} across.}
}
\value{
a \code{data.frame} with entries for \code{name} (the original search term),
\code{guess}, \code{confidence}, and \code{female} and \code{male} with count of matches.
}
\description{
Collect data from the \href{https://www.personality-database.com}{personality-database} API.
}
\examples{
\dontrun{
# get scores from the 5 "HOT"est famous people
hot5 <- personality_database(limit = 5)
}
}
